Fixed some string.h functions: memchr memcmp and memcpy

+++ src/lib/cstdlib/string/memchr.c
@@ -3,7 +3,7 @@
 void* memchr(const void* str, int c, size_t n)
 {
     const char* s = (char*)str;
-    for (int i = 0; i < n; i++)
+    for (size_t i = 0; i < n; i++)
     {
         if (s[i] == c)
             return (void*)&s[i];
